move _delete_scratch_directory(self) to father class JobRunner, and rm it from children class (3 of it)
